🎨 improved code structure
